Add rust binary to docker-compose, add a SQL DB connection to Rust database
Build times are awfully ugly now, sorry for that.
Probably going to have to resort to running on a
non-scratch image with non-static binary or we will
just have to mock the SQL database, or set up separate
docker-compose files. If I figure out how to cache all
of the compiled crates, it should not be such a huge
problem at all