Start implementing repeated scrapping attempts functionality
Started reworking the main program to have more data on the URLs we are
scrapping to allow for more flexibility. It's going to be essentially
a state machine for each url where it goes from Unsent to Sent and
Visited or has to be checked several times because of errors. Had to
switch data structures and tweak the algorithm in the process, but this
apparently sped the program up by a notch.

Still need to implement the actual error handling and repeats, but
this way it's going to be much easier.

Also added some more comments and docs.