Continue implementing repeated scrapping of errored pages
Refactored the code a little more, added more necessary state machine
updates in places and updated the scrapper to allow it to inform the
main thread about errorred scrapping (although error handling itself
in the module is not yet adapted for that).

Currently the implementation is pretty buggy, should work but hold on
till I implement it fully and test it out.