Improve text parsing and look into structured data parsing
Now we should parse full texts a lot better, and I also
escape it now so it'd look much more readable and wouldn't
have to be done later.

Also prepare for transition to a different database, gonna
work on that now. Involves looking into libraries, and
also implementing proper database connection check and retry.

Also I am now looking into why it's so super fast when it starts out
with the URLs it already has and so relatively slow a bit later into
the crawl. Might be because there are not that many new links, might
be because the search into the map slows down because of their sheer
numbers.