Update the README, fix up the code and add comments
I've added some basic instructions on how to launch this app, and also
commented a bit on the code. I understand it's in rough stages, and I
do not intend it as a harsh criticism, just a view from the outside
before you start polishing this app. I've also fixed some of your import
paths and stuff, and just made it so that it would compile and work like
you would (probably) expect it to.

I've not touched the search querying functionality though, it is just
able to compile now and actually shows a home page :)