Unsuccessfully update Dockerfile for rust-crawler
I hate it honestly, I had to switch to a different image and it still
does not work, I've tried a lot of ways to do this, I've spent several
hours staring at my screen hoping that this time I wouldn't have to wait
through the compilation stage, that this time it'd use the cache that it
made already a dozen times. what is currently in here uses the cache if
you do not change any of the files and "rebuilds" it successfully,
otherwise it still downloads several gigs and compiles all of it, once
again. I am obviously missing something